requirements.txt
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
numpy>=1.23
matplotlib>=3.6
scikit-learn>=1.2   
tqdm>=4.65
datasets>=4.1.1         


# === Config & Experiment Management ===
hydra-core>=1.3.2
omegaconf>=2.3.0

mlflow>=2.14.0
tensorboard>=2.17.0
wandb>=0.17.0